third_party/blink/web_tests/external/wpt/css/css-box/margin-trim/computed-margin-values/flexbox-row-block-start-expected.txt

This is a testharness.js-based test.
[FAIL] flexbox > item 1
  assert_equals: \n<item data-expected-margin-top="0" data-offset-y="8"></item>\noffsetTop expected 8 but got 18
[FAIL] flexbox > item 2
  assert_equals: \n<item data-expected-margin-top="0" data-offset-y="8"></item>\noffsetTop expected 8 but got -2
[FAIL] flexbox > item 3
  assert_equals: \n<item data-expected-margin-top="0" data-offset-y="8"></item>\noffsetTop expected 8 but got 71
Harness: the test ran to completion.
